Create hosts copy

+++ refresh_hosts.sh
@@ -1,3 +1,11 @@
 #! /bin/bash
-sudo cp /etc/hosts_original /etc/hosts
+
+#create original copy if not already done
+#otherwise reload oringal version
+if [ -e /etc/hosts_original ] 
+then 
+	sudo cp /etc/hosts_original /etc/hosts
+else
+	sudo cp /etc/hosts /etc/hosts_original
+fi
 sudo cat /home/paul/Dropbox/local_hosts >> /etc/hosts
